requirements.txt
fastapi[all]>=0.109.2  # Latest FastAPI
alembic>=1.13.1  # Latest Alembic
SQLAlchemy>=2.0.27  # Latest SQLAlchemy 2.0
pydantic>=2.6.1  # Latest Pydantic v2
pydantic-settings>=2.1.0  # Required for settings management
python-multipart>=0.0.7  # Form data parsing
prometheus-client>=0.19.0  # Latest metrics client
Pillow>=10.2.0  # Latest Pillow for image processing
qrcode>=7.4.2  # QR code generation
psutil>=5.9.8  # System metrics for health checks
msal>=1.26.0  # Microsoft Authentication Library
python-jose[cryptography]>=3.3.0  # JWT token handling
itsdangerous>=2.1.2  # Session signing

# Test dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
pytest-env>=1.1.1
pytest-xdist>=3.5.0
coverage>=7.4.1
Faker>=22.5.1

# Type checking
types-SQLAlchemy>=1.4.0
types-Pillow>=10.1.0
types-PyYAML>=6.0.0

# Development tools
black>=24.2.0  # Code formatting
ruff>=0.3.0  # Fast Python linter
mypy>=1.8.0  # Static type checking